+++ src/landmark.ml
@@ -256,6 +256,7 @@ type textual_option = {threshold : float}
 type profile_format =
   | JSON
   | Textual of textual_option
+  | External of string
 
 let profiling_ref = ref false
 let profile_with_debug = ref false
@@ -696,6 +697,38 @@ let stop_profiling () =
 
 (** EXPORTING / IMPORTING SLAVE PROFILINGS **)
 
+let warning s =
+  Printf.eprintf "[LANDMARKS] %s.\n%!" s
+
+type exporter = out_channel -> Graph.graph -> unit
+
+let external_exporters : (string, exporter) Hashtbl.t = Hashtbl.create 10
+
+let register_exporter format_name exporter =
+  if Hashtbl.mem external_exporters format_name then
+    warning (Printf.sprintf
+               "Multiple registration of an exporter named %S" format_name);
+  Hashtbl.replace external_exporters format_name exporter
+
+let get_format_names () =
+  Hashtbl.fold (fun name _f acc -> name :: acc)
+    external_exporters
+    ["json"; "textual"]
+  |> List.sort String.compare
+
+let invoke_external_exporter format_name oc graph =
+  match Hashtbl.find_opt external_exporters format_name with
+  | None ->
+      warning (
+        Printf.sprintf
+          "Missing exporter for 'format=%s'. \
+           Available formats are: %s"
+          format_name
+          (String.concat ", " (get_format_names ()))
+      )
+  | Some exporter ->
+      exporter oc graph
+
 let array_list_map f l =
   let size = List.length l in
   match l with
@@ -783,6 +816,8 @@ let exit_hook () =
         Graph.output ~threshold out cg
     | Channel out, JSON ->
         Graph.output_json out cg
+    | Channel out, External name ->
+        invoke_external_exporter name out cg
     | Temporary temp_dir, format ->
         let tmp_file, oc =
           Filename.open_temp_file ?temp_dir "profile_at_exit" ".tmp"
@@ -792,7 +827,8 @@ let exit_hook () =
         flush stdout;
         (match format with
          | Textual {threshold} -> Graph.output ~threshold oc cg
-         | JSON -> Graph.output_json oc cg);
+         | JSON -> Graph.output_json oc cg
+         | External name -> invoke_external_exporter name oc cg);
         close_out oc
   end
 
@@ -809,9 +845,6 @@ let parse_env_options s =
   let allocated_bytes = ref false in
   let split_trim c s =
     List.map String.trim (Misc.split c s)
-  in
-  let warning s =
-    eprintf "[LANDMARKS] %s.\n%!" s
   in
   let parse_option s =
     let invalid_for opt given =
@@ -843,7 +876,9 @@ let parse_env_options s =
         | _ -> format := Textual {threshold = 1.0};
         end
     | [ "format"; "json" ] -> format := JSON;
-    | [ "format"; unknown ] -> invalid_for "format" unknown
+    | [ "format"; other ] ->
+        (* a exporter will have to be registered *)
+        format := External other
     | [ "output"; "stderr" ] -> output := Channel stderr
     | [ "output"; "stdout" ] -> output := Channel stdout
     | [ "output"; temporary ] when Misc.starts_with ~prefix:"temporary" temporary ->
@@ -884,10 +919,13 @@ let parse_env_options s =
   {debug = !debug; allocated_bytes = !allocated_bytes; sys_time = !sys_time;
    output = !output; format = !format; recursive = !recursive}
 
-let () = match Sys.getenv "OCAML_LANDMARKS" with
+let init () =
+  match Sys.getenv "OCAML_LANDMARKS" with
   | exception Not_found -> ()
   | str ->
       try start_profiling ~profiling_options:(parse_env_options str) ()
       with Exit -> ()
+
+let () = init ()
 
 external raise : exn -> 'a = "%raise"